pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "variant"
version = "0.1.0"
description = "AI financial research agent using expectations investing framework"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-groq>=0.2.0",
    "anthropic>=0.40.0",
    "yfinance>=0.2.50",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "tavily-python>=0.5.0",
    "ddgs>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ipykernel",
]

[project.scripts]
variant = "variant.main:main"
